1. 修改菜单图片存储目录 2. 修改数据库结构，将用户数据与其他数据合并 3. 添加拖动版在线调度网页

+++ WebApplication2/Controllers/OrderController.cs
@@ -80,8 +80,8 @@ namespace WebApplication2.Controllers
             MenuOrder menuOrder = db.MenuOrders.Find(id);
 
             menuOrder.Address = Request["address"] == null ? "" : Request["address"];
-            menuOrder.JD = decimal.Parse(Request["JD"]);
-            menuOrder.WD = decimal.Parse(Request["WD"]);
+            menuOrder.JD = double.Parse(Request["JD"]);
+            menuOrder.WD = double.Parse(Request["WD"]);
             menuOrder.Text = Request["text"] == null ? "" : Request["text"];
             db.Entry(menuOrder).State = EntityState.Modified;
             db.SaveChanges();
